Added vector_addressing headers and .cu files instead of calling them utils
